==> analyses/04-01_data-preps-grid.Rmd <==
---
title: "Spatial analyses of 2020-21 excess mortality in CH"
subtitle: "Grid & grided data preparation"
pagetitle: "Spatial Excess: grid"
author: "Radek Panczak, Garyfallos Konstantinoudis & Julien Riou"
date: "`r Sys.Date()`"
output:
  rmdformats::robobook:
    code_folding: show
    self_contained: true
    highlight: pygments
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ----------------------------------------------------- -->

```{r setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(
  tidyverse, lubridate, magrittr, scales, haven,
  DT,
  sf, tmap
)

tmap_mode("view")

import::from("sjmisc", "frq")
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()

conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(
  cache = FALSE,
  prompt = FALSE,
  tidy = FALSE,
  comment = NA,
  message = FALSE,
  warning = FALSE
)

knitr::opts_knit$set(width = 75)
```

<!-- ----------------------------------------------------- -->

# Grid

## Municipality boundaries

Used to 'clip' the data to the country outline, 'remove' the area of lakes and introduce the boundaries within grid cells belonging to two different municipalities.   

```{r}
st_gg21 <- read_rds("data/swisstopo/st_gg21.Rds") %>%
  st_transform(2056)

st_se21_sel <- read_rds("data/swisstopo/st_se21_sel.Rds") %>%
  st_transform(2056)
```

## Offset

<!-- Openly available STATPOP dataset is used to define `offset` (ie. starting point; lowest left corner) for `st_make_grid`.   -->

```{r eval=FALSE, include=FALSE}
origin_statpop <- read_delim("data-raw/BfS/ag-b-00.03-vz2020statpop/STATPOP2020.zip",
                             delim = ";", escape_double = FALSE, trim_ws = TRUE
) %>%
  mutate_all(as.integer) %>%
  summarise(
    minX = min(E_KOORD),
    minY = min(N_KOORD)
  ) %>%
  st_as_sf(
    coords = c("minX", "minY"),
    crs = 2056,
    remove = FALSE
  )

qtm(origin_statpop)
```

```{r eval=FALSE, include=FALSE}
offset_statpop <- read_delim("data-raw/BfS/ag-b-00.03-vz2020statpop/STATPOP2020.zip",
                             delim = ";", escape_double = FALSE, trim_ws = TRUE
) %>%
  mutate_all(as.integer) %>%
  summarise(
    minX = min(E_KOORD),
    minY = min(N_KOORD)
  )
```

One of the meteoswiss datasets is used to define `offset` (ie. starting point; lowest left corner) for `st_make_grid`.  

Using meteoswiss origin allows us to perfectly align gridded mortality and population data to meteo data.   

```{r}
p_load(ncdf4)

TabsD_14 <- nc_open("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201401010000_201412010000.nc")

E <- ncvar_get(TabsD_14, "E") # X, lon
N <- ncvar_get(TabsD_14, "N") # Y, lat

offset_meteoswiss <- as_tibble(cbind(
  minX = as.integer(min(E)),
  minY = as.integer(min(N))
))

p_unload(ncdf4)
```

```{r eval=FALSE, include=FALSE}
# offset_statpop
offset_meteoswiss

origin_meteoswiss <- st_as_sf(tibble(E = min(E), N = min(N)),
                              coords = c("E", "N")
) %>%
  st_set_crs(2056)

qtm(origin_meteoswiss)
```

## Grid creation

1-km grid, with lower left corner defined using `STATPOP` cells.

Grid is then overlaid with *swisstopo municipality data* to select only cells that overlay it.  

At next step the areas of lakes are 'clipped' from the grid.  

Lastly, grid integrating municipality boundaries is also created.

Also, adding ID for each cell (based on row number) and area for each of the iterations.  

```{r eval=FALSE}
grid_1km <- st_gg21 %>%
  st_make_grid(
    cellsize = 1000, square = TRUE,
    offset = c(offset_meteoswiss$minX, offset_meteoswiss$minY)
  ) %>%
  st_sf() %>%
  st_cast("POLYGON") %>%
  st_filter(st_gg21, join = st_covers) %>%
  mutate(ID = row_number()) %>%
  relocate(ID) %>%
  st_transform(2056)

# this one takes some time
grid_1km_crop <- grid_1km %>%
  st_intersection(st_union(st_gg21)) %>%
  st_collection_extract("POLYGON") %>%
  mutate(area = st_area(.)) %>%
  relocate(ID, area) %>%
  st_transform(2056)

grid_1km_over <- grid_1km_crop %>%
  st_intersection(st_gg21) %>%
  mutate(
    ID2 = row_number(),
    area2 = st_area(.)
  ) %>%
  relocate(ID, ID2, area, area2) %>%
  st_transform(2056)
```

```{r eval=FALSE, include=FALSE}
st_drop_geometry(grid_1km_crop) %>% 
  anti_join(st_drop_geometry(grid_1km))

st_drop_geometry(grid_1km) %>% 
  anti_join(st_drop_geometry(grid_1km_crop))

grid_1km %>% 
  st_drop_geometry() %>% 
  group_by(ID) %>% 
  filter(n() > 1)

grid_1km_crop %>% 
  # st_drop_geometry() %>% 
  group_by(ID) %>% 
  filter(n() > 1) %>% 
  arrange(ID)

grid_1km_crop %>% 
  filter(ID == 11794) %>% 
  qtm()

grid_1km_crop %>% 
  filter(ID == 32997) %>% 
  qtm()
```

```{r eval=FALSE, include=FALSE}
# grids
st_write(grid_1km, "data/grid/grid_1km.gpkg", layer = "grid_1km", delete_dsn = TRUE)
st_write(grid_1km_crop, "data/grid/grid_1km.gpkg", layer = "grid_1km_crop", append = TRUE)
st_write(grid_1km_over, "data/grid/grid_1km.gpkg", layer = "grid_1km_over", append = TRUE)

# admin bound
st_write(st_gg21, "data/grid/grid_1km.gpkg", layer = "st_gg21", append = TRUE)
st_write(st_se21_sel, "data/grid/grid_1km.gpkg", layer = "st_se21_sel", append = TRUE)

write_rds(grid_1km, "data/grid/grid_1km.Rds")
write_rds(grid_1km_crop, "data/grid/grid_1km_crop.Rds")
write_rds(grid_1km_over, "data/grid/grid_1km_over.Rds")
```

```{r include=FALSE}
grid_1km <- read_rds("data/grid/grid_1km.Rds")
grid_1km_crop <- read_rds("data/grid/grid_1km_crop.Rds")
grid_1km_over <- read_rds("data/grid/grid_1km_over.Rds")
```

```{r eval=FALSE, include=FALSE}
View(st_drop_geometry(grid_1km))
plot(st_geometry(grid_1km))

View(st_drop_geometry(grid_1km_crop))
plot(st_geometry(grid_1km_crop))

View(st_drop_geometry(grid_1km_over))
plot(st_geometry(grid_1km_over))

View(st_drop_geometry(grid_1km_over))

grid_1km_over %>%
  filter(ID == 15) %>%
  qtm()
```

## Grid example  

```{r echo=FALSE}
tm_shape(st_gg21 %>% filter(GMDNR == 351)) +
  tm_borders(col = "darkorchid") +
  tm_shape(st_join(grid_1km,
                   st_gg21 %>% filter(GMDNR == 351),
                   left = FALSE
  )) +
  tm_borders(col = "red")
```

## Grid comparison

Example of different integration of boundaries of `Gottlieben` municipality.  

Comparing grid of 'regular' cells (green) and cells 'cropped' by outlines of the country and water bodies. Only grid cells that overlay area of municipality are shown (purple):  

```{r echo=FALSE}
tm_shape(st_gg21 %>% filter(GMDNR == 4651)) +
  tm_borders(col = "darkorchid") +
  tm_shape(st_join(grid_1km,
                   st_gg21 %>% filter(GMDNR == 4651),
                   left = FALSE
  )) +
  tm_borders(col = "forestgreen") +
  tm_shape(st_join(grid_1km_crop,
                   st_gg21 %>% filter(GMDNR == 4651),
                   left = FALSE
  )) +
  tm_borders(col = "red")
```

Comparing grid of 'cropped' cells (red again) with grid integrating community boundaries (yellow). Note subdivision of regular cells into smaller ([sometimes very small!](https://github.com/RPanczak/ISPM_geo-mortality/issues/49)) units:   

```{r echo=FALSE}
tm_shape(st_join(grid_1km_crop,
                 st_gg21 %>% filter(GMDNR == 4651),
                 left = FALSE
)) +
  tm_borders(col = "red") +
  tm_shape(st_join(grid_1km_over,
                   st_gg21 %>% filter(GMDNR == 4651),
                   left = FALSE
  )) +
  tm_borders(col = "yellow")
```

## Grid centroids

```{r}
grid_1km_cent <- st_centroid(grid_1km)

grid_1km_cent %<>%
  mutate(
    X = st_coordinates(.)[, 1],
    Y = st_coordinates(.)[, 2]
  ) %>%
  relocate(geometry, .after = last_col())
```

```{r include=FALSE}
grid_1km_crop_cent <- st_centroid(grid_1km_crop)

grid_1km_crop_cent %<>%
  mutate(
    X = st_coordinates(grid_1km_crop_cent)[, 1],
    Y = st_coordinates(grid_1km_crop_cent)[, 2]
  ) %>%
  # select(-area) %>% 
  relocate(geometry, .after = last_col())

st_write(grid_1km_cent, "data/grid/grid_1km.gpkg", layer = "grid_1km_cent", append = FALSE)
st_write(grid_1km_crop_cent, "data/grid/grid_1km.gpkg", layer = "grid_1km_crop_cent", append = FALSE)

write_rds(grid_1km_cent, "data/grid/grid_1km_cent.Rds")
write_rds(grid_1km_crop_cent, "data/grid/grid_1km_crop_cent.Rds")
```

```{r eval=FALSE, include=FALSE}
View(st_drop_geometry(grid_1km))
plot(st_geometry(grid_1km))
```

Example of Bern:    

```{r echo=FALSE}
tm_shape(st_gg21 %>% filter(GMDNR == 351)) +
  tm_borders(col = "darkorchid") +
  tm_shape(st_join(grid_1km,
                   st_gg21 %>% filter(GMDNR == 351),
                   left = FALSE
  )) +
  tm_borders(col = "red") +
  tm_shape(st_join(grid_1km_cent,
                   st_gg21 %>% filter(GMDNR == 351),
                   left = FALSE
  )) +
  tm_dots()
```

Of course different grids will result in slightly different centroids in areas that were modified:  

```{r echo=FALSE}
tm_shape(st_join(grid_1km,
                 st_gg21 %>% filter(GMDNR == 4651),
                 left = FALSE
)) +
  tm_borders(col = "red") +
  tm_shape(st_join(grid_1km_crop,
                   st_gg21 %>% filter(GMDNR == 4651),
                   left = FALSE
  )) +
  tm_borders(col = "forestgreen") +
  tm_shape(st_join(grid_1km_cent,
                   st_join(grid_1km,
                           st_gg21 %>% filter(GMDNR == 4651),
                           left = FALSE
                   ),
                   left = FALSE
  )) +
  tm_dots(col = "red") +
  tm_shape(st_join(grid_1km_crop_cent,
                   st_join(grid_1km,
                           st_gg21 %>% filter(GMDNR == 4651),
                           left = FALSE
                   ),
                   left = FALSE
  )) +
  tm_dots(col = "forestgreen")
```

```{r include=FALSE}
rm(grid_1km_crop, grid_1km_crop_cent); gc()
```

## Grid <> muicipality link

Issue discussed in [#57](https://github.com/RPanczak/ISPM_excess-mortality-spatial/issues/57).  

### Simple spatial link

Assigns the municipality code to the simple geometrical centroid.  

```{r eval=FALSE, include=FALSE}
grid_1km_cent %>% 
  filter(ID == 11794) %>% 
  qtm()

grid_1km_crop_cent %>% 
  filter(ID == 11794) %>% 
  qtm()

grid_1km_cent %>% 
  filter(ID == 32997) %>% 
  qtm()

grid_1km_crop_cent %>% 
  filter(ID == 32997) %>% 
  qtm()
```

```{r}
grid_gemeinde <- grid_1km_cent %>% 
  st_join(st_gg21, join = st_nearest_feature) %>% 
  select(-X, -Y, -border) %>% 
  st_drop_geometry()
```

```{r eval=FALSE, include=FALSE}
grid_1km_cent %>% 
  right_join(
    grid_gemeinde %>% 
      filter(GMDNR == 351)
  ) %>% 
  qtm()
```

```{r include=FALSE}
write_rds(grid_gemeinde, "data/grid/grid_gemeinde.Rds")
```

### Population weighted link

Alternative solution is to use `STATPOP` **2020** dataset prepared in `03`. This solution uses all building coordinates within each grid cell, weighted by population to find alternative centroid.     

```{r}
statpop2020_agg <- read_rds("data/BfS-closed/STATPOP/statpop2020_agg.Rds") %>%
  dplyr::mutate(X = sf::st_coordinates(.)[,1],
                Y = sf::st_coordinates(.)[,2]) %>% 
  relocate(geometry, .after = last_col())

statpop2020_grid <- 
  st_join(statpop2020_agg, 
          grid_1km, 
          join = st_intersects) %>% 
  relocate(ID)

grid_gemeinde_pop_xy <- statpop2020_grid %>% 
  st_drop_geometry() %>% 
  group_by(ID) %>% 
  summarize(X = as.integer(stats::weighted.mean(X, pop)),
            Y = as.integer(stats::weighted.mean(Y, pop))) %>% 
  ungroup() %>% 
  st_as_sf(coords = c("X", "Y"), 
           crs = 2056)

grid_gemeinde_pop <- grid_gemeinde_pop_xy %>% 
  st_join(st_gg21, join = st_nearest_feature) %>% 
  select(-border) %>% 
  st_drop_geometry()
```

Example of cell from Bern agglo area showing example of two definitions of grid centroids which will determine assignments of grid cells to municipality level variables (such as level of urbanization or voting patterns, etc.)

- geometrical centroid (lower left marker) which assigns grid cell to Bremgarten
- population weighted centroid ((upper right marker) (purple circles are buildings scaled to population size) which assigns cell to Bern

```{r echo=FALSE}
cell_number = 26844

tm_shape(st_gg21 %>% filter(GMDNAME == "Bremgarten bei Bern")) + 
  tm_polygons() + 
  tm_shape(grid_1km %>% filter(ID == cell_number)) + 
  tm_borders() + 
  tm_shape(st_filter(statpop2020_agg, grid_1km %>% filter(ID == cell_number))) +
  tm_dots(col = "darkorchid4", shape = 16, size = "pop", alpha = 0.5) + 
  tm_shape(grid_1km_cent %>% filter(ID == cell_number)) + 
  tm_markers(col = "black") + 
  tm_shape(grid_gemeinde_pop_xy %>% filter(ID == cell_number)) + 
  tm_markers(col = "darkorchid4")
```


```{r include=FALSE}
write_rds(grid_gemeinde_pop, "data/grid/grid_gemeinde_pop.Rds")
```

```{r eval=FALSE, include=FALSE}
# most populous grid
statpop2020_grid %>% 
  st_drop_geometry() %>% 
  group_by(ID) %>% 
  summarize(pop_agg = sum(pop)) %>% 
  arrange(desc(pop_agg))

statpop2020_grid %>% 
  filter(ID == 4084) %>% 
  qtm()

# most buildings in grid
statpop2020_grid %>% 
  st_drop_geometry() %>% 
  group_by(ID) %>% 
  summarize(geb_agg = n()) %>% 
  arrange(desc(geb_agg))

statpop2020_grid %>% 
  filter(ID == 40186) %>% 
  qtm()
```

<!-- ----------------------------------------------------- -->

# Gridded Swiss-SEP

```{r}
sep3 <- read_rds("../SNC_Swiss-SEP2/FINAL/RDS/ssep3_user_geo.Rds") %>%
  select(gisid, ssep3, ssep3_d) %>% 
  st_transform(crs = 2056)
```

Using `r number(nrow(sep3), big.mark = ",")` n'hoods from version 3.0:  

```{r echo=FALSE}
sep3 %>% 
  st_drop_geometry() %>% 
  frq(ssep3_d)
```

## Grid based averages

SEP dataset is aggregated to the level of `r scales::number(length(unique(grid_1km$ID)), big.mark = ",")` grid cells.  


```{r eval=FALSE}
ssep3_grid_1km <- 
  st_join(sep3, 
          grid_1km, 
          join = st_intersects) 

ssep3_neigh <- ssep3_grid_1km %>% 
  st_drop_geometry() %>% 
  group_by(ID) %>% 
  summarise(n = n(),
            ssep3_median = median(ssep3)) %>% 
  ungroup() 

ssep3_neigh_grid <- left_join(grid_1km, ssep3_neigh) %>% 
  # filter(!is.na(ssep3_median)) %>% 
  mutate(median_ssep3_d = ntile(ssep3_median, 10)) %>% 
  mutate(median_ssep3_d = factor(median_ssep3_d,
                                 levels = 1:10,
                                 labels = c("1st - lowest", 
                                            "2", "3", "4", 
                                            "5th decile", 
                                            "6", "7", "8", "9", 
                                            "10th - highest"))) %>% 
  relocate(geometry, .after = last_col()) %>% 
  mutate(n = ifelse(is.na(n), 0, n))

write_rds(ssep3_neigh_grid, "data/BfS-closed/SEP/ssep3_neigh_grid.rds")
```

```{r include=FALSE}
ssep3_neigh_grid <- read_rds("data/BfS-closed/SEP/ssep3_neigh_grid.rds")
```

Distribution of number of neighbourhoods (ie. points with SEP) per grid cell:  

```{r echo=FALSE}
ggplot(ssep3_neigh_grid) + 
  geom_histogram(aes(x = n), binwidth = 10)
```

Final map using median index:  

```{r include=FALSE}
tmap_mode("view")
```

```{r eval=FALSE, include=FALSE}
frq(ssep3_neigh_grid$median_ssep3_d)
```

```{r echo=FALSE}
ssep3_neigh_grid %>%
  filter(n > 0) %>% 
  qtm(borders = NULL, fill = "median_ssep3_d", fill.palette = "div")
```

## Population based averages

Alternative solution is to use `STATPOP` **2020** dataset prepared in `03`. This solution uses assigns SEP to nearest building of STATPOP which in turns gives population numbers.   

```{r eval=FALSE}
# STATPOP 2020
statpop2020_agg <- read_rds("data/BfS-closed/STATPOP/statpop2020_agg.Rds")

# join to sep3
ssep3_pop <- st_join(statpop2020_agg, sep3, join = st_nearest_feature)

# takes some time
# nearest <- st_nearest_feature(statpop2020_agg, sep3)
# ssep3_pop$dist3 <- st_distance(ssep3_pop, sep3[nearest, ], by_element = TRUE)
# rm(nearest, statpop2020_agg); gc()
# summary(ssep3_pop$dist3)

ssep3_pop <- 
  st_join(ssep3_pop, grid_1km, join = st_intersects) %>%  
  relocate(egid, gisid, pop,
           ssep3, ssep3_d,
           ID) 

write_rds(ssep3_pop, "data/BfS-closed/SEP/ssep3_pop.Rds")
```

```{r include=FALSE}
rm(statpop2020_agg); gc()

ssep3_pop <- read_rds("data/BfS-closed/SEP/ssep3_pop.Rds") %>% 
  st_drop_geometry()
```

```{r}
ssep3_pop_grid <- ssep3_pop %>% 
  select(ID, ssep3, pop) %>% 
  uncount(pop) %>% 
  group_by(ID) %>% 
  summarise(pop = n(),
            ssep3_median = median(ssep3)) %>% 
  ungroup() 

ssep3_pop_grid <- left_join(grid_1km, ssep3_pop_grid) %>% 
  # filter(!is.na(ssep3_median)) %>% 
  mutate(median_ssep3_d = ntile(ssep3_median, 10)) %>% 
  mutate(median_ssep3_d = factor(median_ssep3_d,
                                 levels = 1:10,
                                 labels = c("1st - lowest", 
                                            "2", "3", "4", 
                                            "5th decile", 
                                            "6", "7", "8", "9", 
                                            "10th - highest"))) %>% 
  relocate(geometry, .after = last_col()) %>% 
  mutate(n = ifelse(is.na(pop), 0, pop))

write_rds(ssep3_pop_grid, "data/BfS-closed/SEP/ssep3_pop_grid.rds")
```

Final map using median index - bubbles scaled to number of people (STATPOP 2020)  

```{r eval=FALSE, include=FALSE}
frq(ssep3_pop_grid$median_ssep3_d)
```

```{r echo=FALSE}
ssep3_pop_grid %>% 
  tm_shape() +
  tm_dots(size = "pop", col = "median_ssep3_d", palette = "RdYlGn",
          scale = .5,  shape = 19)
```

Final map using median index:  

```{r eval=FALSE, include=FALSE}
frq(ssep3_pop_grid$median_ssep3_d)
```

```{r echo=FALSE}
ssep3_pop_grid %>%
  filter(pop > 0) %>% 
  qtm(borders = NULL, fill = "median_ssep3_d", fill.palette = "div")
```

## Comparing two solutions 

### Bland-Altman 1

```{r}
p_load(blandr)

compare <- 
  full_join(
    ssep3_neigh_grid %>% 
      # filter(n > 0) %>% 
      st_drop_geometry() %>% 
      select(ID, ssep3_median) %>% 
      rename(neighbourhood = ssep3_median),
    
    ssep3_pop_grid %>% 
      # filter(pop > 0) %>% 
      st_drop_geometry() %>% 
      select(ID, ssep3_median) %>% 
      rename(population = ssep3_median)
  ) %>% 
  na.omit()

blandr.output.text(compare$neighbourhood, compare$population, sig.level=0.95)
p_unload(blandr)
```

### Bland-Altman 2

```{r}
p_load(BlandAltmanLeh)

bas_list <- bland.altman.stats(compare$neighbourhood, compare$population)
bas <- tibble(means = bas_list$means, diffs = bas_list$diffs)

# bas_list$mean.diffs 
# bas_list$critical.diff 

ggplot(bas, aes(x = means, y = diffs)) + 
  geom_count(alpha = 0.05) + 
  # geom_point() + 
  geom_hline(yintercept = bas_list$lines[-2], 
             linetype = 2, 
             size = 1,
             colour = "slateblue1") + 
  geom_hline(yintercept = bas_list$lines[2], 
             linetype = 2, 
             size = 1, 
             colour = "slateblue4") + 
  xlab("Mean of measurements") + 
  ylab("Difference") + 
  theme_classic()

p_unload(BlandAltmanLeh)
```

Mean diff and CI lines:

```{r}
bas_list$lines
```

Proportion within 95%:

```{r}
prop.table(table(bas_list$diffs >= bas_list$lines[1] & bas_list$diffs <= bas_list$lines[3]))
```

## Decile discrepancies

```{r}
compare <- 
  full_join(
    ssep3_neigh_grid %>% 
      # filter(n > 0) %>% 
      st_drop_geometry() %>% 
      select(ID, median_ssep3_d) %>% 
      rename(neighbourhood = median_ssep3_d),
    
    ssep3_pop_grid %>% 
      # filter(pop > 0) %>% 
      st_drop_geometry() %>% 
      select(ID, median_ssep3_d) %>% 
      rename(population = median_ssep3_d)
  ) %>% 
  na.omit()

# with(compare, table(neighbourhood, population))

start <- nrow(compare)

compare %>% 
  filter(population == neighbourhood) %>% 
  nrow() -> agreement

compare %>% 
  mutate(diff = as.numeric(population) - as.numeric(neighbourhood)) %>% 
  filter(diff >= -1 & diff <= 1) %>% 
  nrow() -> partial

compare %>% 
  mutate(diff = as.numeric(population) - as.numeric(neighbourhood)) %>% 
  filter(diff == -9 | diff == 9) %>% 
  nrow() -> disagreement
```

Out of `r number(start, big.mark = ",")` grid cells `r number(agreement, big.mark = ",")` or `r percent(agreement/start)` has been assigned to the same decile using both solutions. 

If we relax this assumption and alow variation +/- 1 decile then `r percent(partial/start)` of cells are matching.

There are only `r number(disagreement, big.mark = ",")` (`r percent(disagreement/start, accuracy = 0.001)`) cells that are on the opposite scale of the index. 

### Full numbers

Counts

```{r}
sep_dif_xtab <- table(compare$neighbourhood, compare$population) # first var in rows
# sep_dif_xtab

sep_dif_xtab %>% # 
  knitr::kable(caption = 'SSEP 1.0 (rows) vs. 2.0 (columns).')
```

Share (cell %)

```{r}
prop.table(sep_dif_xtab) %>% # cell percentages
  knitr::kable(digits = 3,
               caption = '% SSEP 1.0 (rows) vs. 2.0 (columns).')
```

```{r eval=FALSE, include=FALSE}
mosaicplot(sep_dif_xtab, shade = TRUE)

spineplot(factor(compare$neighbourhood), factor(compare$population))
```

Visually 

```{r, layout="l-page", fig.width=6, fig.height=6}
sep_dif_freq <- as.data.frame(table(compare$neighbourhood, compare$population)) %>% 
  rename(SSEP1 = Var1, SSEP2 = Var2)

sep_dif_freq %>% 
  ggplot(aes(SSEP1, Freq)) +
  geom_col() +
  facet_grid(vars(SSEP2))
```

Alternative view

```{r, layout="l-body-outset"}
sep_dif_freq %>% 
  ggplot(aes(SSEP1, SSEP2)) +
  geom_point(aes(size = Freq, color = Freq / nrow(compare)), shape = 15) +
  scale_size_continuous(range = c(2, 10)) + 
  scale_color_distiller(palette = "RdYlBu", direction = 1) +
  theme_classic()
```

Excluding diagonal

```{r, layout="l-body-outset"}
sep_dif_freq %>% 
  filter(SSEP1 != SSEP2) %>% 
  ggplot(aes(SSEP1, SSEP2)) +
  geom_point(aes(size = Freq, color = Freq / nrow(compare)), shape = 15) +
  scale_size_continuous(range = c(2, 10)) + 
  scale_color_distiller(palette = "RdYlBu", direction = 1) +
  theme_classic()
```

<!-- ----------------------------------------------------- -->

# Gridded temperature data

Using gridded data from [meteoswiss](https://www.meteoswiss.admin.ch/home/climate/swiss-climate-in-detail/raeumliche-klimaanalysen.html).  

Using `TabsM`: *Monthly mean temperature* dataset for the years of interest.  

## Monthly data 

Example of two months in 2014:  

```{r echo=FALSE}
p_load(raster)

filename <- "data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201401010000_201412010000.nc"

# r <- raster::stack(filename) # same same
r <- raster::brick(filename)
raster::crs(r) <- st_crs(2056)$proj4string

# nlayers(r)

plot(r[[c(1, 7)]],
     main = "Monthly temps 2014 - January & July",
     zlim = c(-20, 25),
     col = rev(RColorBrewer::brewer.pal(11, "RdBu"))
)

rm(filename, r)
```

**Note:** *there is known, unsolved issue with projection of meteoswiss' nc files!*  

```{r eval=FALSE, include=FALSE}
plot(r[[1]],
     main = "TabsM_14 - January",
     zlim = c(-20, 25),
     col = rev(RColorBrewer::brewer.pal(11, "RdBu"))
)

plot(st_geometry(grid_1km_cent), add = TRUE)
```

## Extract

Extract function  

```{r eval=FALSE}
extract_temps <- function(filename, year) {
  
  # r <- raster::stack(filename) # same same
  r <- raster::brick(filename)
  raster::crs(r) <- sf::st_crs(2056)$proj4string
  
  ex <- raster::extract(r, grid_1km_cent, df = TRUE)
  ex <- tibble::as_tibble(ex)
  repeats <- nrow(ex)
  
  ex <- pivot_longer(ex,
                     !ID,
                     names_to = "date_temp",
                     values_to = "temperature"
  )
  
  ex$year <- as.integer(year)
  ex$month <- rep(seq(1, 12), repeats)
  ex$date_temp <- NULL
  
  return(ex)
}
```

Combining years:  

```{r eval=FALSE}
temperature <- bind_rows(
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201401010000_201412010000.nc", 2014),
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201501010000_201512010000.nc", 2015),
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201601010000_201612010000.nc", 2016),
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201701010000_201712010000.nc", 2017),
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201801010000_201812010000.nc", 2018),
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_201901010000_201912010000.nc", 2019),
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_202001010000_202012010000.nc", 2020),
  extract_temps("data-raw/meteoswiss/TabsM_ch01r.swiss.lv95_202101010000_202112010000.nc", 2021)
) %>%
  filter(!(year == 2021 & month > 6))
```

```{r eval=FALSE, include=FALSE}
write_rds(temperature, "data/meteoswiss/temperature.Rds")
```

```{r include=FALSE}
p_unload(raster)
temperature <- read_rds("data/meteoswiss/temperature.Rds")
```

## Examples of time series

Data for Lugano & Jungfraujoch.  

```{r echo=FALSE}
temp_checks <-
  tibble(
    Place = c("Jungfraujoch", "Lugano"),
    lon = c(7.980556, 8.9525),
    lat = c(46.547222, 46.005)
  ) %>%
  st_as_sf(
    coords = c("lon", "lat"),
    crs = 4326,
    remove = FALSE
  ) %>%
  st_transform(2056)

# qtm(temp_checks)

temp_checks %<>%
  st_join(grid_1km) %>%
  st_drop_geometry() %>%
  select(Place, ID) %>%
  left_join(temperature) %>%
  select(-ID) %>%
  mutate(Date = ym(paste(year, month, sep = " ")))

ggplot(temp_checks, aes(Date, temperature,
                        group = Place, col = Place
)) +
  geom_line() +
  geom_point() +
  theme_light() +
  ylab("Temperature")
```

```{r include=FALSE}
rm(temp_checks); gc()
```

<!-- ----------------------------------------------------- -->

# Gridded deaths 2014-2019

## SNC data 

Prepared in `/Stata/mort_2014-2020.do`.  

```{r}
snc5_deaths <- read_dta("data-raw/BfS-closed/SNC/deaths.dta") %>%
  zap_formats() %>%
  zap_labels() %>%
  zap_label() %>%
  dplyr::select(-geoyear, -dob) %>%
  dplyr::rename(year = yod) %>%
  mutate(
    yod = as.integer(year),
    sex = if_else(sex == "1", "Female", "Male"),
    sex = factor(sex),
    age = cut(age_death,
              breaks = c(0, 40, seq(50, 80, 10), 120), right = FALSE,
              labels = c("<40", "40-49", "50-59", "60-69", "70-79", "80+")
    ),
    month = month(dod)
  ) %>%
  relocate(month, .after = year) %>%
  dplyr::select(-age_death, -dod) %>%
  relocate(age, .after = sex) %>%
  st_as_sf(
    coords = c("geox", "geoy"),
    crs = 2056,
    remove = TRUE
  )
```

```{r eval=FALSE, include=FALSE}
snc5_deaths %>%
  st_drop_geometry() %>%
  group_by(age) %>%
  summarise(
    min = min(year),
    max = max(year)
  )

snc5_deaths %>%
  sample_n(100) %>%
  qtm()

View(st_drop_geometry(snc5_deaths))

cell_number = 26844

grid_1km_over %>% 
  filter(ID == cell_number) %>% 
  qtm(text = "ID2")
```

## Aggregate deaths by grid X municipality

```{r eval=FALSE}
temp1 <- st_join(snc5_deaths,
                 grid_1km_over,
                 join = st_intersects
) %>%
  st_drop_geometry() %>%
  group_by(ID, ID2, GMDNR, GMDNAME, year, month, sex, age) %>%
  summarize(deaths = as.integer(n())) %>%
  ungroup() 

temp2 <- temp1 %>% 
  select(ID, ID2, GMDNR, GMDNAME) %>% 
  distinct() %>% 
  expand_grid(
    year = unique(temp1$year),
    month = unique(temp1$month),
    sex = unique(temp1$sex),
    age = unique(temp1$age)
  ) %>%
  dplyr::arrange(year, month, sex, age) 

st_grid_deaths_2014_2019 <- 
  temp2 %>% 
  left_join(temp1) %>% 
  mutate(deaths = replace_na(deaths, 0)) %>%
  mutate(deaths = as.integer(deaths)) %>% 
  dplyr::arrange(ID, ID2, GMDNR, GMDNAME, year, month, sex, age) %>% 
  relocate(ID, ID2, GMDNR, GMDNAME)
```

```{r eval=FALSE, include=FALSE}
rm(temp1, temp2, snc5_deaths); gc()

write_rds(st_grid_deaths_2014_2019, "data/blob/st_grid_deaths_2014_2019.Rds")

cell_number = 26844

grid_1km_over %>% 
  filter(ID == cell_number) %>% 
  left_join(st_grid_deaths_2014_2019) %>% 
  filter(sex == "Female" & age == "80+",
         year == 2014 & month == 1)
```

```{r include=FALSE}
st_grid_deaths_2014_2019 <- read_rds("data/blob/st_grid_deaths_2014_2019.Rds")
```

<!-- ----------------------------------------------------- -->

# Gridded population data 2014-2020

## Function 

Defined in `geom_001_pe_to_grid.R`.  

```{r, eval=FALSE}
source("R/geom_001_pe_to_grid.R")
```

## Sources 

STATPOP data, state of 31st Dec each year. Aggregating individual level data to building level, spatial join to grid.  

Example of most recent data:  

```{r, eval=FALSE}
gc()
r20_pe_grid <- geom_001_pe_to_grid("data-raw/BfS-closed/STATPOP/statpop2020_220098p.zip")
write_rds(r20_pe_grid, "data/BfS-closed/STATPOP/r20_pe_grid.Rds")
rm(r20_pe_grid)
gc()
```

```{r, eval=FALSE, include=FALSE}
gc()
r19_pe_grid <- geom_001_pe_to_grid("data-raw/BfS-closed/STATPOP/statpop2019_220098p.zip")
write_rds(r19_pe_grid, "data/BfS-closed/STATPOP/r19_pe_grid.Rds")
rm(r19_pe_grid)
gc()

gc()
r18_pe_grid <- geom_001_pe_to_grid("data-raw/BfS-closed/STATPOP/statpop2018_220098p.zip")
write_rds(r18_pe_grid, "data/BfS-closed/STATPOP/r18_pe_grid.Rds")
rm(r18_pe_grid)
gc()

gc()
r17_pe_grid <- geom_001_pe_to_grid("data-raw/BfS-closed/STATPOP/statpop2017_220098p.zip")
write_rds(r17_pe_grid, "data/BfS-closed/STATPOP/r17_pe_grid.Rds")
rm(r17_pe_grid)
gc()

gc()
r16_pe_grid <- geom_001_pe_to_grid("data-raw/BfS-closed/STATPOP/statpop2016_220098p.zip")
write_rds(r16_pe_grid, "data/BfS-closed/STATPOP/r16_pe_grid.Rds")
rm(r16_pe_grid)
gc()

gc()
r15_pe_grid <- geom_001_pe_to_grid("data-raw/BfS-closed/STATPOP/statpop2015_220098p.zip")
write_rds(r15_pe_grid, "data/BfS-closed/STATPOP/r15_pe_grid.Rds")
rm(r15_pe_grid)
gc()

gc()
r14_pe_grid <- geom_001_pe_to_grid("data-raw/BfS-closed/STATPOP/statpop2014_220098p.zip")
write_rds(r14_pe_grid, "data/BfS-closed/STATPOP/r14_pe_grid.Rds")
rm(r14_pe_grid)
gc()
```

Combining together:  

```{r eval=FALSE}
pe_grid <- bind_rows(
  read_rds("data/BfS-closed/STATPOP/r20_pe_grid.Rds") %>%
    st_drop_geometry() %>% select(-egid),
  read_rds("data/BfS-closed/STATPOP/r19_pe_grid.Rds") %>%
    st_drop_geometry() %>% select(-egid),
  read_rds("data/BfS-closed/STATPOP/r18_pe_grid.Rds") %>%
    st_drop_geometry() %>% select(-egid),
  read_rds("data/BfS-closed/STATPOP/r17_pe_grid.Rds") %>%
    st_drop_geometry() %>% select(-egid),
  read_rds("data/BfS-closed/STATPOP/r16_pe_grid.Rds") %>%
    st_drop_geometry() %>% select(-egid),
  read_rds("data/BfS-closed/STATPOP/r15_pe_grid.Rds") %>%
    st_drop_geometry() %>% select(-egid),
  read_rds("data/BfS-closed/STATPOP/r14_pe_grid.Rds") %>%
    st_drop_geometry() %>% select(-egid)
)
```

## Aggregate by grid/municipality

```{r, eval=FALSE}
lu_grid_gem <- pe_grid %>%
  group_by(year, sex, age, ID, ID2, area, area2, GMDNR, GMDNAME, KTNR, KTNAME) %>%
  summarise(pop = sum(n), .groups = "drop") %>%
  dplyr::arrange(year, ID, ID2, sex, age)
```

## Expand grid for combinations with 0

```{r, eval=FALSE}
lu_grid_gem <- grid_1km_over %>%
  sf::st_drop_geometry() %>%
  dplyr::as_tibble() %>%
  dplyr::select(ID, ID2, area, area2, GMDNR, GMDNAME, KTNR, KTNAME) %>%
  tidyr::expand_grid(
    year = unique(lu_grid_gem$year),
    age = unique(lu_grid_gem$age),
    sex = unique(lu_grid_gem$sex)
  ) %>%
  dplyr::arrange(year, ID, ID2, sex, age) %>%
  dplyr::left_join(lu_grid_gem) %>%
  dplyr::mutate(pop = if_else(is.na(pop), as.integer(0), pop)) %>%
  dplyr::mutate(year = as.integer(year))
```

```{r eval=FALSE, include=FALSE}
lu_grid_gem %>%
  filter(GMDNR == 4651, age == "<40", sex == "Female", year == 2018) %>%
  view()

grid_1km_over %>%
  right_join(
    lu_grid_gem %>%
      filter(GMDNR == 4651, age == "<40", sex == "Female", year == 2018)
  ) %>%
  qtm()

gc()
write_rds(lu_grid_gem, "data/blob/lu_grid_gem.Rds")
```

## Aggregate by grid

```{r, eval=FALSE}
st_grid_pop <- lu_grid_gem %>%
  dplyr::group_by(ID, year, age, sex) %>%
  dplyr::summarise(pop = sum(pop), .groups = "drop")
```

```{r eval=FALSE, include=FALSE}
gc()
write_rds(st_grid_pop, "data/blob/st_grid_pop.Rds")
```

```{r eval=FALSE, include=FALSE}
# sanity check
st_grid_pop_exp %>% filter(year == 2020) %>% summarise(pop = sum(pop))
```

```{r include=FALSE}
st_grid_pop <- read_rds("data/blob/st_grid_pop.Rds")
```

<!-- ----------------------------------------------------- -->

# Gridded deaths 2020-2021

Redistributing municipality level deaths for 2020-2021 into grid.  

Done using `04-02_grid-deaths.R` script on HPC.  

<!-- ----------------------------------------------------- -->

# Extrapolated population after 2019

## Data for modelling

```{r}
st_grid_pop_exp <- dplyr::bind_rows(
  st_grid_pop %>%
    dplyr::filter(year != 2020),
  st_grid_pop %>%
    dplyr::filter(year >= 2018) %>%
    dplyr::mutate(year = year + 2) %>% 
    dplyr::mutate(pop = NA_integer_)
) %>%
  dplyr::arrange(ID, year, age, sex) %>%
  dplyr::mutate(
    # year_month = year + ,
    id_year = year - 2013,
    id_age = as.integer(as.factor(age))
  )
```

```{r include=FALSE}
write_rds(st_grid_pop_exp, "data/blob/st_grid_pop_exp.Rds")
```

```{r eval=FALSE, include=FALSE}
table(st_grid_pop$year)
table(st_grid_pop_exp$year)

length(unique(st_grid_pop$ID))
length(unique(st_grid_pop_exp$ID))
```

## Examples 

Zero pops example:  

```{r echo=FALSE}
filter(st_grid_pop_exp, ID == 1, age == "<40", sex == "Female") %>%
  ggplot(aes(x = year, y = pop)) +
  geom_point()
```

Max pops example:  

```{r echo=FALSE}
filter(st_grid_pop_exp, ID == 4084, age == "<40", sex == "Female") %>%
  ggplot(aes(x = year, y = pop)) +
  geom_point()
```

## Zeroes 

Frequency:  

```{r echo=FALSE}
st_grid_pop_exp %>%
  filter(!is.na(pop)) %>%
  frq(pop == 0)
```

Distribution:  

```{r echo=FALSE}
st_grid_pop_exp %>%
  filter(!is.na(pop)) %>%
  ggplot(aes(x = pop)) +
  geom_histogram(binwidth = 10)
```

## Processing 

Tested using `04-03_grid-pops.R` & `04-04_grid-pops.R` script on HPC.
